services/ingest_worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["provider", "normalize", "load", "dq", "config.py", "main.py"]

[project]
name = "ingest-worker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1.0",
    "google-cloud-storage>=2.14,<3.0",
    "google-cloud-bigquery>=3.17,<4.0",
    "google-cloud-firestore>=2.14,<3.0",
    "structlog>=24.1,<25.0",
    "python-dotenv>=1.0,<2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "respx>=0.21,<1.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
